refactor: add system for multiple parsers
